refactor(extension): rename IP whitelist to allowlist
Use inclusive allowlist terminology for the module, provider, UI labels,
docs, and realm attribute setting keys.
