+++ extensions/ip-allowlist/src/main/resources/META-INF/services/io.github.mabartos.spi.evaluator.RiskEvaluatorFactory
@@ -1,0 +1,1 @@
+io.github.mabartos.evaluator.ip.IpAllowlistRiskEvaluatorFactory
