--- extensions/ip-whitelist/src/main/resources/META-INF/services/io.github.mabartos.spi.evaluator.RiskEvaluatorFactory
@@ -1,1 +1,0 @@
-io.github.mabartos.evaluator.ip.IpWhitelistRiskEvaluatorFactory
