+++ extensions/pom.xml
@@ -12,7 +12,7 @@
     <packaging>pom</packaging>
     <modules>
         <module>ip-api</module>
-        <module>ip-whitelist</module>
+        <module>ip-allowlist</module>
         <module>openrouter</module>
     </modules>
